Display poster in correct aspect ratio and simplify code
Previously, the poster size was set to the video player size, resulting in a distorted poster if sizes didn't match. Now, the poster is shown in its correct size, but the content resizes therefore "skips" by the difference between video and poster heights. When setting the wrapper height to the video height the general page structure does not change, but the poster still changes as it is displayed vertically centered while the player is loading, and then it skips from the vertical center to the top after the error. There are css solutions but they would blow up this demo and make it unnecessarily complicated. SInce the poster is displayed during player initialization anyway, it could just be kept on screen after an error and render this whole demo case obsolete (bitmovin/bitdash#911).